add: intent launcher
You can _almost_ join from your browser now.

+++ src/container/run.nim
@@ -7,12 +7,22 @@ import
 import ../argparser
 import ./utils/[exec, mount]
 
-proc showUI*() =
+proc showUI*(launch: bool = true) =
   var platform = getIPlatformService()
-  platform.launchApp("com.roblox.client")
+  if launch:
+    platform.launchApp("com.roblox.client")
+
   platform.setProperty("waydroid.active_apps", "com.roblox.client")
 
-proc startAndroidRuntime*(input: Input) =
+proc startRobloxClient*(platform: var IPlatform) =
+  while isEmptyOrWhitespace(&readOutput("pidof", "com.roblox.client")):
+    # FIXME: please don't do this
+    platform.launchApp("com.roblox.client")
+    sleep(100)
+
+  startLogWatcher()
+
+proc startAndroidRuntime*(input: Input, launchRoblox: bool = true) =
   info "equinox: starting android runtime"
   debug "equinox: starting prep for android runtime"
 
@@ -29,7 +39,6 @@ proc startAndroidRuntime*(input: Input) =
   settings.fflags["FFlagUserFyosDetectionHorseFly"] = newJBool(true)
 
   setFflags(settings.fflags)
-
   generateSessionLxcConfig()
 
   if getLxcStatus() == "RUNNING":
@@ -40,27 +49,38 @@ proc startAndroidRuntime*(input: Input) =
 
     var platform = getIPlatformService()
     platform.setProperty("waydroid.active_apps", "com.roblox.client")
+    
+    if launchRoblox:
+      startRobloxClient(platform)
 
-    while isEmptyOrWhitespace(&readOutput("pidof", "com.roblox.client")):
-      # FIXME: please don't do this
-      platform.launchApp("com.roblox.client")
-      sleep(100)
+      let pid = parseUint(&readOutput("pidof", "com.roblox.client"))
+      debug "equinox: waiting for roblox to exit: pid=" & $pid
+      while kill(Pid(pid), 0) == 0 or errno != ESRCH:
+        sleep(100)
 
-    startLogWatcher()
-
-    let pid = parseUint(&readOutput("pidof", "com.roblox.client"))
-
-    debug "equinox: waiting for roblox to exit: pid=" & $pid
-    var status: cint
-    while kill(Pid(pid), 0) == 0 or errno != ESRCH:
-      sleep(100)
-
-    if WIFEXITED(status):
-      info "equinox: runtime has been stopped."
-    else:
-      warn "equinox: runtime stopped abnormally."
-
-    stopLogWatcher()
-    stopNetworkService()
-    stopLxcContainer()
+      stopLogWatcher()
+      stopNetworkService()
+      stopLxcContainer()
     # deinitHardwareService(hwsvc)
+
+type
+  PlaceURI* = distinct string
+
+proc launchRobloxGame*(input: Input, id: PlaceURI | string) =
+  info "equinox: sending VIEW intent to roblox client"
+
+  if getLxcStatus() != "RUNNING":
+    startAndroidRuntime(input, launchRoblox = false)
+  else:
+    showUI(launch = false)
+
+  var platform = getIPlatformService()
+
+  while isEmptyOrWhitespace(&readOutput("pidof", "com.roblox.client")):
+    # FIXME: please don't do this
+    when id is string:
+      platform.launchIntent("android.intent.action.VIEW", "roblox://placeId=" & id)
+    else:
+      platform.launchIntent("android.intent.action.VIEW", cast[string](id))
+
+  platform.setProperty("waydroid.active_apps", "com.roblox.client")
